Splits user inputs into separate component
